requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
dj-database-url==1.0.0
dj-email-url==1.0.6
Django==4.1.3
django-cache-url==3.4.2
djangorestframework==3.14.0
environs==9.5.0
flake8==5.0.4
gunicorn==20.1.0
idna==3.4
marshmallow==3.19.0
mccabe==0.7.0
packaging==21.3
psycopg2-binary==2.9.5
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
sqlparse==0.4.3
stripe==5.0.0
urllib3==1.26.12